DataNeo4jTestConfiguration moved to test src folder
